+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added pre-bundled minified versions to each plugin package
 - Added new convenience package `piral-hooks-utils`
 - Added new npm script running bundler `piral-cli-xbuild` (#470)
+- Added support for `--public-url` also in `pilet` commands
 - Added types to `piral-cli-webpack/extend-config` and `piral-cli-webpack5/extend-config`
 - Added `--feed` option to `piral debug` command
 - Added semver possibilities for resolving shared dependencies